LayoutTests/fast/dom/Element/mozilla-dom-base-tests/test_bug1075702.xhtml

<!DOCTYPE html>
<html xml:lang="en" xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1075702
-->
<head>
    <meta charset="utf-8"/>
    <script src="../../../../resources/js-test-pre.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1075702"> Mozilla Bug 1075702 </a>
<p id="display"></p>

<pre id="test"></pre>
<script type="application/javascript">

  /** Test for Bug 1075702 **/
  // test: Element.removeAttributeNode()
  description("Test adapted from Mozilla Bug 1075702");

  var a1 = document.createAttribute("aa");
  a1.nodeValue = "lowercase";

  var a2 = document.createAttribute("AA");
  a2.nodeValue = "UPPERCASE";

  document.documentElement.setAttributeNode(a1);

  document.documentElement.setAttributeNode(a2);

  shouldBeTrue('document.documentElement.hasAttributeNS("", "aa")');
  shouldBeEqualToString('document.documentElement.getAttributeNS("", "aa")', 'lowercase');
  shouldBeTrue('document.documentElement.hasAttributeNS("", "AA")');
  shouldBeEqualToString('document.documentElement.getAttributeNS("", "AA")', 'UPPERCASE');

</script>
<script src="../../../../resources/js-test-post.js"></script>
</body>
</html>
