LayoutTests/fast/dom/Element/mozilla-dom-base-tests/test_bug1075702-expected.txt

Test adapted from Mozilla Bug 1075702

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.documentElement.hasAttributeNS("", "aa") is true
PASS document.documentElement.getAttributeNS("", "aa") is "lowercase"
PASS document.documentElement.hasAttributeNS("", "AA") is true
PASS document.documentElement.getAttributeNS("", "AA") is "UPPERCASE"
PASS successfullyParsed is true

TEST COMPLETE
Mozilla Bug 1075702
